feat: migrate testcases to ts
