+++ package.json
@@ -56,6 +56,7 @@
     "@semantic-release/commit-analyzer": "^9.0.1",
     "@semantic-release/npm": "^8.0.0",
     "@semantic-release/release-notes-generator": "^10.0.2",
+    "@types/jest": "^26.0.1",
     "all-contributors-cli": "^6.17.4",
     "babel-eslint": "^10.1.0",
     "babel-jest": "^26.1.0",
@@ -71,17 +72,21 @@
     "prettier": "^2.0.5",
     "react-addons-test-utils": "^15.6.2",
     "semantic-release": "^18.0.0",
+    "ts-jest": "^26.0.1",
+    "ts-loader": "^8.3.0",
+    "typescript": "^4.4.4",
     "webpack": "^4.43.0",
     "webpack-cli": "^3.3.12"
   },
   "jest": {
+    "preset": "ts-jest/presets/js-with-babel",
     "coverageReporters": [
       "json-summary",
       "text",
       "lcov"
     ],
     "collectCoverageFrom": [
-      "src/**/*.{js,jsx}",
+      "src/**/*.{js,jsx,tx,tsx}",
       "!**/test/**"
     ]
   },
