--- src/js/index.js
@@ -1,25 +1,0 @@
-import {
-  fraudPreventionHeadersEnum,
-  getFraudPreventionHeaders,
-  getScreenDetails,
-  windowDetails,
-  getGovClientBrowserJSUserAgentHeader,
-  getGovClientBrowserPluginsHeader,
-  getGovClientBrowserDoNotTrackHeader,
-  getGovClientDeviceID,
-  getGovClientTimezoneHeader,
-  getGovClientLocalIPsHeader,
-} from "./hmrc/mtdFraudPrevention";
-
-exports.fraudPreventionHeadersEnum = fraudPreventionHeadersEnum;
-exports.getFraudPreventionHeaders = getFraudPreventionHeaders;
-exports.getGovClientBrowserPluginsHeader = getGovClientBrowserPluginsHeader;
-exports.getGovClientBrowserDoNotTrackHeader = getGovClientBrowserDoNotTrackHeader;
-
-exports.getGovClientDeviceID = getGovClientDeviceID;
-exports.getGovClientTimezoneHeader = getGovClientTimezoneHeader;
-
-exports.getGovClientLocalIPsHeader = getGovClientLocalIPsHeader;
-exports.getScreenDetails = getScreenDetails;
-exports.getGovClientBrowserJSUserAgentHeader = getGovClientBrowserJSUserAgentHeader;
-exports.windowDetails = windowDetails;
