+++ src/js/index.ts
@@ -1,0 +1,12 @@
+export {
+  fraudPreventionHeadersEnum,
+  getFraudPreventionHeaders,
+  getScreenDetails,
+  windowDetails,
+  getGovClientBrowserJSUserAgentHeader,
+  getGovClientBrowserPluginsHeader,
+  getGovClientBrowserDoNotTrackHeader,
+  getGovClientDeviceID,
+  getGovClientTimezoneHeader,
+  getGovClientLocalIPsHeader,
+} from './hmrc/mtdFraudPrevention';
